+++ lib/routes/get-refund/widgets/refund_item.dart
@@ -1,4 +1,4 @@
-import 'package:breez_sdk/bridge_generated.dart';
+import 'package:breez_sdk/sdk.dart';
 import 'package:c_breez/routes/get-refund/widgets/refund_item_action.dart';
 import 'package:c_breez/widgets/available_btc.dart';
 import 'package:flutter/material.dart';
@@ -15,7 +15,7 @@ class RefundItem extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       children: [
-        AvailableBTC(swapInfo.confirmedSats),
+        AvailableBTC(swapInfo.confirmedSats.toInt()),
         RefundItemAction(swapInfo),
       ],
     );
